[Wip] search config...
