--- tapi-front/store/getters.js
@@ -1,5 +1,0 @@
-export default {
-  isSignedIn(state) {
-    return state.status
-  }
-};
