--- tapi-front/store/state.js
@@ -1,10 +1,0 @@
-export default {
-  // user relation
-  user: null,         // user name
-  user_token: null,   // user token
-  status: false,      // user status flag
-
-  // search form relation
-  lists: [],          // response API data
-  isLoading: false    // loading flag
-};
